M2kCalibration: Drop the pimpl implementation and use pure abstract class.
